+++ test/ConfigurationSpec.scala
@@ -124,11 +124,11 @@ class ConfigurationSpec extends FlatSpec with ShouldMatchers with DefaultConvert
     config5[String]("nums.five") should be ("V")
   }
 
-  it can "be dettach from a configuration" in {
+  it can "be detach from a configuration" in {
     val data2 = Map( "one" -> "1", "two" -> "2" )
     val config2 = Configuration( data2 )
     val config3 = config attach ("nums", config2 )
-    val config4 = config3 dettach ("nums")
+    val config4 = config3 detach ("nums")
     config4 should be (config2)
   }
 
@@ -180,8 +180,7 @@ class ConfigurationObjectSpec extends FlatSpec with ShouldMatchers with io.IOHel
     config.get[String]("baz") should be (Some("hello world"))
   }
 
-  it can "be loaded from a file" in {
-    val filename = "/tmp/configrity_configuration_obj_spec.conf"
+  it can "be loaded from a file" in {    
     val fmt = FlatFormat
     val s = 
       """
@@ -189,7 +188,7 @@ class ConfigurationObjectSpec extends FlatSpec with ShouldMatchers with io.IOHel
       bar = 2
       baz = "hello world"
       """
-    autoFile( filename, s ){ file =>
+    autoFile( s ){ file =>
       val fn = file.getAbsolutePath
       val config = Configuration.load(fn,fmt)
       config.get[Boolean]("foo") should be (Some(true))